Added EventSender trait and an implementation managing multiple event listeners.
There are multiple kinds of event receivers in the JGUIraffe environment.
Some of them are really listeners which can be registered at a UI component.
A specific event listener is the event manager which handles multiplexing of
standard events. The EventSender trait was introduced to handle these different
listeners in a generic way.