Added a generic IAnswer implementation for retrieving a specific parameter object.
In unit tests it happens from time to time that a method invocation is expected
with an object created dynamically during test execution time. This object is
important for the test case. The new FetchAnswer class allows handling such
cases in a generic way. It has already been integrated into the
TestMouseEventAdapter class.