Changed the handling of the window's root container.
A Java FX window now manages a Group object as its root container. This group
is populated during the build process. When the window is opened it is
added to the wrapped stage's scene. This is necessary because of the
threading restrictions in Java FX. Scene objects must not be manipulated
outside the Java FX application thread. This also includes manipulations
of the nodes in the scene's root object.