Added an initial implementation of TableColumnRecalibrationResizePolicy.
This is a callback invoked by a table view when a column resize operation is
performed. This implementation mainly triggers the column width controller to
perform a recalibration. Due to problems of the standard JavaFX resize
policy implementations, the resize operation is also checked and implemented
here.

The current implementation is incomplete. It does not yet react on auto-fit
operations triggered by a user by a double click on the table header.