TableColumnRecalibrationResizePolicy can now react on independent width changes.
If the user performs an auto-fit on a column, the resize policy is not
informed. Therefore, we can add change listeners on the widths of the
columns of the table in question. However, these listeners also fire after
resize operations handled by the policy. So the policy must be aware of this
and ignore a certain number of change events.