TableViewWidthChangeListener now takes the table's border into account.
When calculating the new column widths the whole width of the table cannot be
used. Rather, a value for the table's border has to be subtracted; otherwise,
a horizontal scrollbar is displayed. Unfortunately, there does not seem to be
an easy way to determine the exact border width value. Therefore, the value
is empirical.

A related discussion can also be found at
https://community.oracle.com/thread/2377826?tstart=0