Changed the way the current cell is handled by CellComponentManager.
Whenever a new row becomes the current row of the table, the whole render form
is initialized. So the approach taken before - to set the current form when a
cell selects itself - failed because at that time the cell was not yet
processed, and therefore, the current form was undefined.

In the new approach, proxies for component handlers pass their data to the
CellComponentManager. From there it is propagated to render cells when they
are selected.