Fixed problems in the handling of model changes in JavaFxTreeHandler.
Obviously, adding new nodes to a tree node causes the new nodes to be selected
directly. This causes issues when the model is dynamically updated, e.g. if
data is loaded and added to the tree in reaction to user clicks. Therefore,
during a resync operation, no more selection change notifications are
generated. The handling of the selection change property has been reworked.