Updated changes.xml and changes report.
The report did not generate correct URLs pointing to the referenced issues.
It does not seem to be easily possible to support multiple trackers (for bugs
and feature requests) at the same time. Therefore, now full URLs for issues
have been added. This is not perfect either because now no links are generated
(only the URLs are output as plain text).