[44]: Made access to Swing tool tips more consistent.
When setting the tool tip of a Swing component the text was converted
to HTML, but the HTML markup was not removed when the tool tip was
queried. So a different tool tip text was returned when the tip was
queried than the tip that was set. This has been changed.