Build: Reorganized plugin management.
The versions of plugins are now defined as properties and referenced
from the `pluginManagement` section. This also includes the
`scala-maven-plugin` only used by the JavaFX module.

In order to make the `site:site` target work, the version of the
`spotbugs` plugin had to be increased; the version that was used
before could not process classes from a newer Java version. With these
changes, the `site:site` goal is now successful.

Signed-off-by: Oliver Heger <oliver.heger@oliver-heger.de>